chore: rename `constants` into `modes`
